pharmacovigilance-demo/cloud_ingestion/requirements.txt
azure-functions
azure-eventhub
requests
